Add a security group and a prefix list
The security group provides fro all the communication necessary
between the VENOM agents and the VENOM CIDR blocks.  The prefix list
is a convenient way to manage the multiple VENOM CIDR blocks and
reference them in the security group rules.

Note that we also must add a policy attachment that allows us to
create these resources.